feat: refactor frontend into separate files (#131)
* feat: refactor frontend code into separate files

* chore: lint
